`CoreDataFeedStore.retrieve()` delivers found values on non-empty cache
